requirements.txt
# Rasa is optional for local development
# rasa>=3.6.0 # Requires Python 3.8-3.10
rasa-sdk>=3.6.0
requests>=2.28.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
responses>=0.23.3
#tenacity>=9.1.2
tenacity>=8.0.0